[Var] Vorbis: Add libvorbis 1.3.5. [Mod] libopenmpt: Use libvorbisfile instead of stb_vorbis in the default MSVC builds in order to the default configuration in sync with Linux builds.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@6073 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ include/vorbis/include/vorbis/vorbisfile.h
@@ -55,10 +55,10 @@ static int _ov_header_fseek_wrap(FILE *f,ogg_int64_t off,int whence){
 
 #ifdef __MINGW32__
   return fseeko64(f,off,whence);
-#elif ( defined(_WIN32) && !defined(_MSC_VER) ) || ( defined(_WIN32) && defined(_MSC_VER) && _MSC_VER >= 1400 )
+#elif defined (_WIN32)
   return _fseeki64(f,off,whence);
 #else
-  return fseek(f,(long)off,whence);
+  return fseek(f,off,whence);
 #endif
 }
